refactor(workflows): consolidate safeSendMessage into executor-shared
executor.ts and dag-executor.ts each defined a near-identical copy of
safeSendMessage and SendMessageContext. Move both into executor-shared.ts
as exports and import from there. The UnknownErrorTracker guard is also
relocated and exported, but remains dormant infrastructure: no callsite
in either file instantiates a tracker today, so this PR delivers pure
deduplication with no behavioral change. Activating the guard in DAG
execution is appropriate follow-up work.

Also addresses two issues surfaced during code review (#1496):
- Reset UnknownErrorTracker.count on any non-UNKNOWN error (was only
  resetting on success), so a sequence like UNKNOWN -> TRANSIENT ->
  UNKNOWN -> UNKNOWN no longer trips the threshold incorrectly.
- Use the structured 'platform.critical_message_send_failed' event name
  in the inlined log call instead of the prose 'Critical message send
  failed' string.
