+++ packages/workflows/src/executor-shared.ts
@@ -7,7 +7,7 @@
  */
 import { readFile } from 'fs/promises';
 import { join } from 'path';
-import type { WorkflowDeps } from './deps';
+import type { IWorkflowPlatform, WorkflowDeps, WorkflowMessageMetadata } from './deps';
 import * as archonPaths from '@archon/paths';
 import { BUNDLED_COMMANDS, isBinaryBuild } from './defaults/bundled-defaults';
 import { createLogger } from '@archon/paths';
@@ -531,4 +531,83 @@ export function stripCompletionTags(content: string, until?: string): string {
  */
 export function isInlineScript(script: string): boolean {
   return script.includes('\n') || /[;(){}&|<>$`"' ]/.test(script);
+}
+
+// ─── Platform Message Sending ────────────────────────────────────────────────
+
+/** Context for platform message sending */
+export interface SendMessageContext {
+  workflowId?: string;
+  nodeName?: string;
+}
+
+/** Threshold for consecutive UNKNOWN errors before aborting */
+const UNKNOWN_ERROR_THRESHOLD = 3;
+
+/** Mutable counter for tracking consecutive unknown errors across calls */
+export interface UnknownErrorTracker {
+  count: number;
+}
+
+/**
+ * Safely send a message to the platform without crashing on failure.
+ * Returns true if message was sent successfully, false otherwise.
+ * Only suppresses transient/unknown errors; fatal errors are rethrown.
+ * When unknownErrorTracker is provided, consecutive UNKNOWN errors are tracked
+ * and the workflow is aborted after UNKNOWN_ERROR_THRESHOLD consecutive failures.
+ */
+export async function safeSendMessage(
+  platform: IWorkflowPlatform,
+  conversationId: string,
+  message: string,
+  context?: SendMessageContext,
+  metadata?: WorkflowMessageMetadata,
+  unknownErrorTracker?: UnknownErrorTracker
+): Promise<boolean> {
+  try {
+    await platform.sendMessage(conversationId, message, metadata);
+    if (unknownErrorTracker) unknownErrorTracker.count = 0;
+    return true;
+  } catch (error) {
+    const err = error as Error;
+    const errorType = classifyError(err);
+
+    getLog().error(
+      {
+        err,
+        conversationId,
+        messageLength: message.length,
+        errorType,
+        platformType: platform.getPlatformType(),
+        ...context,
+        stack: err.stack,
+      },
+      'platform_message_send_failed'
+    );
+
+    // Reset tracker on any non-UNKNOWN outcome — only *consecutive* UNKNOWN
+    // errors should trip the threshold (e.g. UNKNOWN→TRANSIENT→UNKNOWN→UNKNOWN
+    // is two separate runs, not three in a row).
+    if (unknownErrorTracker && errorType !== 'UNKNOWN') {
+      unknownErrorTracker.count = 0;
+    }
+
+    // Fatal errors should not be suppressed - they indicate configuration issues
+    if (errorType === 'FATAL') {
+      throw new Error(`Platform authentication/permission error: ${err.message}`);
+    }
+
+    // Track consecutive UNKNOWN errors - abort if threshold exceeded
+    if (errorType === 'UNKNOWN' && unknownErrorTracker) {
+      unknownErrorTracker.count++;
+      if (unknownErrorTracker.count >= UNKNOWN_ERROR_THRESHOLD) {
+        throw new Error(
+          `${String(UNKNOWN_ERROR_THRESHOLD)} consecutive unrecognized errors - aborting workflow: ${err.message}`
+        );
+      }
+    }
+
+    // Transient errors (and below-threshold unknown errors) suppressed to allow workflow to continue
+    return false;
+  }
 }
